finalized readme and added screenshots
