--- test-DONT-TOUCH/test.css
@@ -1,3 +1,0 @@
-article {
-  margin-left: 30px;
-}
